complete features

+++ Weatherapp/app/src/main/java/com/venki/weatherapp/weatherapp/helpers/Helper.java
@@ -31,8 +31,8 @@ public class Helper {
     }
 
     public static long convertCelsiusToFahrenheit (long centigrade) {
-        long fahrenheit;
-        fahrenheit = centigrade * (9/5) + 32;
-        return fahrenheit;
+        double fahrenheit;
+        fahrenheit = centigrade * ((float)9/(float)5) + 32;
+        return (long)fahrenheit;
     }
 }
